# code-template/target/classes/template/source/swagger/com/changgou/swagger/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "微服务工程代码生成器,生成Dao、Service、Controller、Pojo、Feign、Swagger的JSON脚本",
    "version": "1.0.0",
    "title": "深圳黑马训练营代码生成器",
    "termsOfService": "https://github.com/shenkunlin/code-template.git",
    "contact": {
      "email": "shenkunlin@itcast.cn"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "host": "petstore.swagger.io",
  "basePath": "/v2",
...
  "schemes": [
    "http"
  ],
  "paths": {
...
    "/oauthClientDetails/{id}": {
      "delete": {
        "tags": [
          "OauthClientDetailsController"
        ],
        "summary": "根据ID删除OauthClientDetails",
        "description": "根据ID删除OauthClientDetails方法详情",
        "operationId": "deleteUsingDELETE",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "description": "根据ID删除OauthClientDetails方法详情",
            "required": true
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid status value(无效的状态值)"
                      },
          "403": {
            "description": "403 Forbidden(请求被拒绝)"
                      },
          "404": {
            "description": "not found(没有找到相关资源)"
                      },
          "405": {
            "description": "Invalid input(无效的输入)"
                      },
          "500": {
            "description": "服务器内部错误"
                      },
          "200": {
            "description": "根据ID删除OauthClientDetails"
            ,
            "schema": {
              "$ref": "#/definitions/Result"
            }
          }
        }
      },
      "put": {
        "tags": [
          "OauthClientDetailsController"
        ],
        "summary": "根据ID修改OauthClientDetails",
        "description": "根据ID修改OauthClientDetails方法详情",
        "operationId": "updateUsingPUT",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "description": "根据ID修改OauthClientDetails方法详情",
            "required": true
          },
          {
            "in": "body",
            "name": "OauthClientDetails",
            "description": "传入OauthClientDetails的JSON对象",
            "required": false
              ,
              "schema": {
                "$ref": "#/definitions/OauthClientDetails"
              }
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid status value(无效的状态值)"
                      },
          "403": {
            "description": "403 Forbidden(请求被拒绝)"
                      },
          "404": {
            "description": "not found(没有找到相关资源)"
                      },
          "405": {
            "description": "Invalid input(无效的输入)"
                      },
          "500": {
            "description": "服务器内部错误"
                      },
          "200": {
            "description": "根据ID修改OauthClientDetails"
            ,
            "schema": {
              "$ref": "#/definitions/Result"
            }
          }
        }
      },
      "get": {
        "tags": [
          "OauthClientDetailsController"
        ],
        "summary": "根据ID查询OauthClientDetails",
        "description": "根据ID查询OauthClientDetails方法详情",
        "operationId": "findByIdUsingGET",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "description": "根据ID修改OauthClientDetails方法详情",
            "required": true
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid status value(无效的状态值)"
                      },
          "403": {
            "description": "403 Forbidden(请求被拒绝)"
                      },
          "404": {
            "description": "not found(没有找到相关资源)"
                      },
          "405": {
            "description": "Invalid input(无效的输入)"
                      },
          "500": {
            "description": "服务器内部错误"
                      },
          "200": {
            "description": "根据ID修改OauthClientDetails"
            ,
            "schema": {
              "$ref": "#/definitions/Result«OauthClientDetails»"
            }
          }
        }
      }
    },
    "/oauthClientDetails/search": {
      "post": {
        "tags": [
          "OauthClientDetailsController"
        ],
        "summary": "不带分页条件搜索OauthClientDetails",
        "description": "不带分页条件搜索OauthClientDetails方法详情",
        "operationId": "findListUsingPOST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "OauthClientDetails",
            "description": "传入OauthClientDetails的JSON对象",
            "required": false
              ,
              "schema": {
                "$ref": "#/definitions/OauthClientDetails"
              }
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid status value(无效的状态值)"
                      },
          "403": {
            "description": "403 Forbidden(请求被拒绝)"
                      },
          "404": {
            "description": "not found(没有找到相关资源)"
                      },
          "405": {
            "description": "Invalid input(无效的输入)"
                      },
          "500": {
            "description": "服务器内部错误"
                      },
          "200": {
            "description": "不带分页搜索OauthClientDetails"
            ,
            "schema": {
              "$ref": "#/definitions/Result«List«OauthClientDetails»»"
            }
          }
        }
      }
    },
    "/oauthClientDetails/search/{page}/{size}": {
      "get": {
        "tags": [
          "OauthClientDetailsController"
        ],
        "summary": "分页列表查询OauthClientDetails",
        "description": "分页列表查询OauthClientDetails方法详情",
        "operationId": "findPageUsingGET",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "page",
            "description": "当前页",
            "required": true
          },
          {
            "in": "path",
            "name": "size",
            "description": "每页显示条数",
            "required": true
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid status value(无效的状态值)"
                      },
          "403": {
            "description": "403 Forbidden(请求被拒绝)"
                      },
          "404": {
            "description": "not found(没有找到相关资源)"
                      },
          "405": {
            "description": "Invalid input(无效的输入)"
                      },
          "500": {
            "description": "服务器内部错误"
                      },
          "200": {
            "description": "不带分页搜索OauthClientDetails"
            ,
            "schema": {
              "$ref": "#/definitions/Result«List«OauthClientDetails»»"
            }
          }
        }
      },
      "post": {
        "tags": [
          "OauthClientDetailsController"
        ],
        "summary": "分页条件搜索OauthClientDetails",
        "description": "分页条件搜索OauthClientDetails方法详情",
        "operationId": "findPageUsingPOST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "page",
            "description": "当前页",
            "required": true
          },
          {
            "in": "path",
            "name": "size",
            "description": "每页显示条数",
            "required": true
          },
          {
            "in": "body",
            "name": "OauthClientDetails",
            "description": "传入OauthClientDetails的JSON对象",
            "required": false
              ,
              "schema": {
                "$ref": "#/definitions/OauthClientDetails"
              }
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid status value(无效的状态值)"
                      },
          "403": {
            "description": "403 Forbidden(请求被拒绝)"
                      },
          "404": {
            "description": "not found(没有找到相关资源)"
                      },
          "405": {
            "description": "Invalid input(无效的输入)"
                      },
          "500": {
            "description": "服务器内部错误"
                      },
          "200": {
            "description": "分页条件搜索OauthClientDetails"
            ,
            "schema": {
              "$ref": "#/definitions/Result«List«OauthClientDetails»»"
            }
          }
        }
      }
    },
...
    "/areas/{id}": {
      "delete": {
        "tags": [
          "AreasController"
        ],
        "summary": "根据ID删除Areas",
        "description": "根据ID删除Areas方法详情",
        "operationId": "deleteUsingDELETE",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "description": "根据ID删除Areas方法详情",
            "required": true
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid status value(无效的状态值)"
                      },
          "403": {
            "description": "403 Forbidden(请求被拒绝)"
                      },
          "404": {
            "description": "not found(没有找到相关资源)"
                      },
          "405": {
            "description": "Invalid input(无效的输入)"
                      },
          "500": {
            "description": "服务器内部错误"
                      },
          "200": {
            "description": "根据ID删除Areas"
            ,
            "schema": {
              "$ref": "#/definitions/Result"
            }
          }
        }
      },
      "put": {
        "tags": [
          "AreasController"
        ],
        "summary": "根据ID修改Areas",
        "description": "根据ID修改Areas方法详情",
        "operationId": "updateUsingPUT",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "description": "根据ID修改Areas方法详情",
            "required": true
          },
          {
            "in": "body",
            "name": "Areas",
            "description": "传入Areas的JSON对象",
            "required": false
              ,
              "schema": {
                "$ref": "#/definitions/Areas"
              }
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid status value(无效的状态值)"
                      },
          "403": {
            "description": "403 Forbidden(请求被拒绝)"
                      },
          "404": {
            "description": "not found(没有找到相关资源)"
                      },
          "405": {
            "description": "Invalid input(无效的输入)"
                      },
          "500": {
            "description": "服务器内部错误"
                      },
          "200": {
            "description": "根据ID修改Areas"
            ,
            "schema": {
              "$ref": "#/definitions/Result"
            }
          }
        }
      },
      "get": {
        "tags": [
          "AreasController"
        ],
        "summary": "根据ID查询Areas",
        "description": "根据ID查询Areas方法详情",
        "operationId": "findByIdUsingGET",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "description": "根据ID修改Areas方法详情",
            "required": true
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid status value(无效的状态值)"
                      },
          "403": {
            "description": "403 Forbidden(请求被拒绝)"
                      },
          "404": {
            "description": "not found(没有找到相关资源)"
                      },
          "405": {
            "description": "Invalid input(无效的输入)"
                      },
          "500": {
            "description": "服务器内部错误"
                      },
          "200": {
            "description": "根据ID修改Areas"
            ,
            "schema": {
              "$ref": "#/definitions/Result«Areas»"
            }
          }
        }
      }
    },
    "/areas/search": {
      "post": {
        "tags": [
          "AreasController"
        ],
        "summary": "不带分页条件搜索Areas",
        "description": "不带分页条件搜索Areas方法详情",
        "operationId": "findListUsingPOST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "Areas",
            "description": "传入Areas的JSON对象",
            "required": false
              ,
              "schema": {
                "$ref": "#/definitions/Areas"
              }
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid status value(无效的状态值)"
                      },
          "403": {
            "description": "403 Forbidden(请求被拒绝)"
                      },
          "404": {
            "description": "not found(没有找到相关资源)"
                      },
          "405": {
            "description": "Invalid input(无效的输入)"
                      },
          "500": {
            "description": "服务器内部错误"
                      },
          "200": {
            "description": "不带分页搜索Areas"
            ,
            "schema": {
              "$ref": "#/definitions/Result«List«Areas»»"
            }
          }
        }
      }
    },
    "/areas/search/{page}/{size}": {
      "get": {
        "tags": [
          "AreasController"
        ],
        "summary": "分页列表查询Areas",
        "description": "分页列表查询Areas方法详情",
        "operationId": "findPageUsingGET",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "page",
            "description": "当前页",
            "required": true
          },
          {
            "in": "path",
            "name": "size",
            "description": "每页显示条数",
            "required": true
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid status value(无效的状态值)"
                      },
          "403": {
            "description": "403 Forbidden(请求被拒绝)"
                      },
          "404": {
            "description": "not found(没有找到相关资源)"
                      },
          "405": {
            "description": "Invalid input(无效的输入)"
                      },
          "500": {
            "description": "服务器内部错误"
                      },
          "200": {
            "description": "不带分页搜索Areas"
            ,
            "schema": {
              "$ref": "#/definitions/Result«List«Areas»»"
            }
          }
        }
      },
      "post": {
        "tags": [
          "AreasController"
        ],
        "summary": "分页条件搜索Areas",
        "description": "分页条件搜索Areas方法详情",
        "operationId": "findPageUsingPOST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "page",
            "description": "当前页",
            "required": true
          },
          {
            "in": "path",
            "name": "size",
            "description": "每页显示条数",
            "required": true
          },
          {
            "in": "body",
            "name": "Areas",
            "description": "传入Areas的JSON对象",
            "required": false
              ,
              "schema": {
                "$ref": "#/definitions/Areas"
              }
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid status value(无效的状态值)"
                      },
          "403": {
            "description": "403 Forbidden(请求被拒绝)"
                      },
          "404": {
            "description": "not found(没有找到相关资源)"
                      },
          "405": {
            "description": "Invalid input(无效的输入)"
                      },
          "500": {
            "description": "服务器内部错误"
                      },
          "200": {
            "description": "分页条件搜索Areas"
            ,
            "schema": {
              "$ref": "#/definitions/Result«List«Areas»»"
            }
          }
        }
      }
    },
...
    "/cities/{id}": {
      "delete": {
        "tags": [
          "CitiesController"
        ],
        "summary": "根据ID删除Cities",
        "description": "根据ID删除Cities方法详情",
        "operationId": "deleteUsingDELETE",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "description": "根据ID删除Cities方法详情",
            "required": true
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid status value(无效的状态值)"
                      },
          "403": {
            "description": "403 Forbidden(请求被拒绝)"
                      },
          "404": {
            "description": "not found(没有找到相关资源)"
                      },
          "405": {
            "description": "Invalid input(无效的输入)"
                      },
          "500": {
            "description": "服务器内部错误"
                      },
          "200": {
            "description": "根据ID删除Cities"
            ,
            "schema": {
              "$ref": "#/definitions/Result"
            }
          }
        }
      },
      "put": {
        "tags": [
          "CitiesController"
        ],
        "summary": "根据ID修改Cities",
        "description": "根据ID修改Cities方法详情",
        "operationId": "updateUsingPUT",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "description": "根据ID修改Cities方法详情",
            "required": true
          },
          {
            "in": "body",
            "name": "Cities",
            "description": "传入Cities的JSON对象",
            "required": false
              ,
              "schema": {
                "$ref": "#/definitions/Cities"
              }
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid status value(无效的状态值)"
                      },
          "403": {
            "description": "403 Forbidden(请求被拒绝)"
                      },
          "404": {
            "description": "not found(没有找到相关资源)"
                      },
          "405": {
            "description": "Invalid input(无效的输入)"
                      },
          "500": {
            "description": "服务器内部错误"
                      },
          "200": {
            "description": "根据ID修改Cities"
            ,
            "schema": {
              "$ref": "#/definitions/Result"
            }
          }
        }
      },
      "get": {
        "tags": [
          "CitiesController"
        ],
        "summary": "根据ID查询Cities",
        "description": "根据ID查询Cities方法详情",
        "operationId": "findByIdUsingGET",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "description": "根据ID修改Cities方法详情",
            "required": true
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid status value(无效的状态值)"
                      },
          "403": {
            "description": "403 Forbidden(请求被拒绝)"
                      },
          "404": {
            "description": "not found(没有找到相关资源)"
                      },
          "405": {
            "description": "Invalid input(无效的输入)"
                      },
          "500": {
            "description": "服务器内部错误"
                      },
          "200": {
            "description": "根据ID修改Cities"
            ,
            "schema": {
              "$ref": "#/definitions/Result«Cities»"
            }
          }
        }
      }
    },
    "/cities/search": {
      "post": {
        "tags": [
          "CitiesController"
        ],
        "summary": "不带分页条件搜索Cities",
        "description": "不带分页条件搜索Cities方法详情",
        "operationId": "findListUsingPOST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "Cities",
            "description": "传入Cities的JSON对象",
            "required": false
              ,
              "schema": {
                "$ref": "#/definitions/Cities"
              }
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid status value(无效的状态值)"
                      },
          "403": {
            "description": "403 Forbidden(请求被拒绝)"
                      },
          "404": {
            "description": "not found(没有找到相关资源)"
                      },
          "405": {
            "description": "Invalid input(无效的输入)"
                      },
          "500": {
            "description": "服务器内部错误"
                      },
          "200": {
            "description": "不带分页搜索Cities"
            ,
            "schema": {
              "$ref": "#/definitions/Result«List«Cities»»"
            }
          }
        }
      }
    },
    "/cities/search/{page}/{size}": {
      "get": {
        "tags": [
          "CitiesController"
        ],
        "summary": "分页列表查询Cities",
        "description": "分页列表查询Cities方法详情",
        "operationId": "findPageUsingGET",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "page",
            "description": "当前页",
            "required": true
          },
          {
            "in": "path",
            "name": "size",
            "description": "每页显示条数",
            "required": true
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid status value(无效的状态值)"
                      },
          "403": {
            "description": "403 Forbidden(请求被拒绝)"
                      },
          "404": {
            "description": "not found(没有找到相关资源)"
                      },
          "405": {
            "description": "Invalid input(无效的输入)"
                      },
          "500": {
            "description": "服务器内部错误"
                      },
          "200": {
            "description": "不带分页搜索Cities"
            ,
            "schema": {
              "$ref": "#/definitions/Result«List«Cities»»"
            }
          }
        }
      },
      "post": {
        "tags": [
          "CitiesController"
        ],
        "summary": "分页条件搜索Cities",
        "description": "分页条件搜索Cities方法详情",
        "operationId": "findPageUsingPOST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "page",
            "description": "当前页",
            "required": true
          },
          {
            "in": "path",
            "name": "size",
            "description": "每页显示条数",
            "required": true
          },
          {
            "in": "body",
            "name": "Cities",
            "description": "传入Cities的JSON对象",
            "required": false
              ,
              "schema": {
                "$ref": "#/definitions/Cities"
              }
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid status value(无效的状态值)"
                      },
          "403": {
            "description": "403 Forbidden(请求被拒绝)"
                      },
          "404": {
            "description": "not found(没有找到相关资源)"
                      },
          "405": {
            "description": "Invalid input(无效的输入)"
                      },
          "500": {
            "description": "服务器内部错误"
                      },
          "200": {
            "description": "分页条件搜索Cities"
            ,
            "schema": {
              "$ref": "#/definitions/Result«List«Cities»»"
            }
          }
        }
      }
    },
...
    "/provinces/search": {
      "post": {
        "tags": [
          "ProvincesController"
        ],
        "summary": "不带分页条件搜索Provinces",
        "description": "不带分页条件搜索Provinces方法详情",
        "operationId": "findListUsingPOST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "Provinces",
            "description": "传入Provinces的JSON对象",
            "required": false
              ,
              "schema": {
                "$ref": "#/definitions/Provinces"
              }
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid status value(无效的状态值)"
                      },
          "403": {
            "description": "403 Forbidden(请求被拒绝)"
                      },
          "404": {
            "description": "not found(没有找到相关资源)"
                      },
          "405": {
            "description": "Invalid input(无效的输入)"
                      },
          "500": {
            "description": "服务器内部错误"
                      },
          "200": {
            "description": "不带分页搜索Provinces"
            ,
            "schema": {
              "$ref": "#/definitions/Result«List«Provinces»»"
            }
          }
        }
      }
    },
...
    "/user/search": {
      "post": {
        "tags": [
          "UserController"
        ],
        "summary": "不带分页条件搜索User",
        "description": "不带分页条件搜索User方法详情",
        "operationId": "findListUsingPOST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "User",
            "description": "传入User的JSON对象",
            "required": false
              ,
              "schema": {
                "$ref": "#/definitions/User"
              }
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid status value(无效的状态值)"
                      },
          "403": {
            "description": "403 Forbidden(请求被拒绝)"
                      },
          "404": {
            "description": "not found(没有找到相关资源)"
                      },
          "405": {
            "description": "Invalid input(无效的输入)"
                      },
          "500": {
            "description": "服务器内部错误"
                      },
          "200": {
            "description": "不带分页搜索User"
            ,
            "schema": {
              "$ref": "#/definitions/Result«List«User»»"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
    "Result": {
      "type": "object",
      "required": [
        "code",
        "data",
        "flag",
        "message"
      ],
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32",
          "description": "返回状态码,20000:成功,20001:失败,20002:用户名或密码错误,20003:权限不足,20004:远程调用失败,20005:重复操作,20006:没有对应的数据"
        },
        "data": {
          "type": "object",
          "description": "逻辑数据"
        },
        "flag": {
          "type": "boolean",
          "example": false,
          "description": "执行是否成功,true:成功,false:失败"
        },
        "message": {
          "type": "string",
          "description": "提示信息"
        }
      },
      "description": "Result"
    },
...
    "OauthClientDetails": {
      "type": "object",
      "properties": {
        "clientId": {
          "type": "string",
          "description": "客户端ID，主要用于标识对应的应用"
        },
        "resourceIds": {
          "type": "string",
          "description": ""
        },
        "clientSecret": {
          "type": "string",
          "description": "客户端秘钥，BCryptPasswordEncoder加密"
        },
        "scope": {
          "type": "string",
          "description": "对应的范围"
        },
        "authorizedGrantTypes": {
          "type": "string",
          "description": "认证模式"
        },
        "webServerRedirectUri": {
          "type": "string",
          "description": "认证后重定向地址"
        },
        "authorities": {
          "type": "string",
          "description": ""
        },
        "accessTokenValidity": {
          "type": "integer",
          "format": "int32",
          "description": "令牌有效期"
        },
        "refreshTokenValidity": {
          "type": "integer",
          "format": "int32",
          "description": "令牌刷新周期"
        },
        "additionalInformation": {
          "type": "string",
          "description": ""
        },
        "autoapprove": {
          "type": "string",
          "description": ""
        }
      },
      "description": "OauthClientDetails"
    },
...
  },
...
}